Save squeeze
Debian Squeeze is EOL, but we still want to build packages on it.
The main problem is that the repositories were taken down from the
mirrors. Normally we'd use archive for this case, but it seems that
squeeze-updates and squeeze-lts are not (yet?) in it.

So the only solution I found is to pull packages from a snapshot
repository.